Initiative tracker application.
This uses some pre-set players and allows you to play through rounds
of combat, dealing damage to other participants. It does not allow
you to yet apply conditions to another player, although the UI is
in place. When a player is reduced to 0 hit points, they begin to
die and the UI switches to instead track death saving throws.