Start with action, turns track targets fixes #3
In order to keep the model in sync with the dropdown, we need to make the
Turn model object track who the list of targets are (so that it can
pre-select one of them)

Thus, the TurnOptions component needs to be able to handle no turn being
selected. The App dispatches the new START_ENCOUNTER action upon
rendering, so that we can select the first player (and the first set of
targets).