Basic Healing #1
When a combatant is created, capture its initial hp as the max hp.
Duplicate all of the damage actions for healing, but namespace them with a
damage/healing prefix. This affects the turn model too, which now has a
'healingTarget' and a 'damageTarget', as well as 'damage' and 'healing'.

Currently not supported is the ability to heal thyself.