Decouples initiative order
Each combatant has an initiative score. Upon receiving the START_ENCOUNTER
action, this is sorted and an order array is produced. This array contains
ids, so each combatant now has a unique id.

For determining whose current turn it is, we still use an index, but it's
an index into the orders array.

When it comes to selecting targets and applying the damage, we still use
copies of the objects, but we have to do a bit of a dance, doing
combatants.get(order.get(index)), or vice versa. Perhaps the reselect
library can help here.