Create Encounter container
This might feel like one step forward two steps back, but this is a step
in creating Encounters a re-usable component. In the future it may be
worth re-separating out these into containers, but for now the intention
is that it should be easy to switch out top-level containers for when
there is no encounter in progress.