Encounter Builder UI
The encounter builder UI lets the user input enemies, to join the
hardcoded player in an encounter. Currently they can specify name, hp and
initiatives. There is no validation here, which remains as a todo.

START_ENCOUNTER is now no longer dispatched by default, it is dispatched
from the Start button in the encounter builder. This action now also
passes through the enemies data from the builder reducer, which is then
combined into the combatants map in the encounter reducer.

There is now a top-level App container component, which conditionally
renders the Encounter or EncounterBuilder container depending upon
application state.