Encounters now mapped from id -> encounterModel
The root level reducer now contains an encounters sub-reducer, which
itself uses the existing encounter reducer as a sub-reducer. This is so
that encounters can now be uniquely identified within a map, for future
development where we can manage and run more than one encounter.

A lot of actions have had to change, becuase now their payload contains an
encounterId. Within a container, both mapStateToProps and
mapDispatchToProps are additionally passed in a 'ownProps' property, which
is the properties passed by its owner. In this case, we create a
hard-coded encounter id and pass it in so everything works.