Adds home screen
The home screen shows an overview of encounters, plus a list of players.
For the players, we're still using pre-canned players, and there's a form
to add a new player. It's unstyled and the button to add is disabled.

The add encounter button works, firing an ADD_ENCOUNTER button. This
generates an id, which creates a new encounter. We now no longer need the
pre-generated encounter! There's a new encounter status, 'building', which
represents when we're adding enemies to an encounter. The start encounter
button will eventually be placed in the home screen, next to an encounter.