Start an encounter using redux-thunk
This one confused me for a while. But in order to start an encounter, we
need a set of players, and a set of enemies. At the moment a selector
simply brings in all enemies (because they're not yet keyed against a
specific encounter), and another selector returns all players. A selector
based on those returns an object containing both.

Reselectors are available to action creators, provided you use a thunk. So
the START_ENCOUNTER is the first and only thunk, which gets data from the
selector and dispatches the action.

This is a start towards a slight switch of architecture - about doing more
in the action creators and making the reducers more dumb.