Uses redux-form to add players
Redux Form is a higher-order component that manages the state of forms for
you, through the use of a custom reducer. All we need to do is handle the
onSubmit, passed as a prop to the component. When called it dispatches
ADD_PLAYER, which generates an id and transforms the form data into a
model object.

We now no longer have to pre-generate our players!