Assigns enemies to specific encounter. Fixes #8
An id for an encounter is generated when the user presses the Add button,
and it's passed to the EncounterBuilder container as an ownProp. That id
is passed to the ADD_ENEMY action and added as a property on the enemy
object. It's therefore also used on a corresponding selector,
getEnemiesForEncounter, to return all the enemies that are relevant to
this encounter.

Additionally, the getEncounterCombatants selector no longer simply returns
all enemies, it uses our new selector plus the encounter id it's already
supplied.

This might not ultimately be the best data structure, but for now it will
suffice.