Rolling for initiative menu - Fixes #11
Players and enemies have an initiative bonus. For an encounter, they
should be using a rolled d20 plus this bonus to determine order. This
commit introduces a new encounter state - 'initiatives', which when
present will display the Initiatives menu. A user can enter the initiatve
value for each of the combatants and see the current total.

Continuing will dispatch a new BEGIN_COMBAT action, whose job it is to set
up a turn. This is all stuff that was previously being done in
START_ENCOUNTER.