Set hp on players post-encounter #9
A basic implementation. In the action creator, obtain the players data
(players are now an immutable Record) and add them to a results object.
The players reducer now also handles CLOSE_ENCOUNTER and updates the
players map with the hp value for each player that took part in the
encounter.