Adds monsters from 5e SRD
Monsters are stored as a big JSON object, but when loaded they raise an
action, so that in the future we can potentially handle this
asynchronously.

There's now no need for an enemies reducer - as an enemy created during
the encounter builder is an 'instance' of a monster type. The resulting
data item itself refers back to the monsters reducer so that further
information can be pulled from that store, rather than duplicated. But
this effectively becomes part of the encounter state now.

Several selectors have become redundant because of the state being better
co-located.