Generate enemy names during building
This one's a bit weird and a bit lovely. Rather than using
redux-forms/CHANGE action, we pass in a thunk so that when an enemy type
is selected during encounter building, we can dispatch two actions. One is
to alter the value of the select field, the other is to set the name
field.