Moving container components into their module folders
Rather than having the connected components in a containers/ folder,
prefer a more modular structure where the connected component and its
underlying components are all together.

For reasonably simple components both the connected component and the
underlying component are in the same index.js file, whereas in complex
cases they're split and required as before.

In future we may also add the styles into these modules and import them as
appropriate - which may be a good opportunity for CSS modules.